Task 5 finnished (No unit tests yet)

+++ models/engine/file_storage.py
@@ -1,20 +1,23 @@
 #!/usr/bin/python3
 """
-file_storage module - Represents the definition of 
+file_storage module - Represents the definition of
 the class FileStorage that serializes instances
 to a JSON file and deserializes JSON file to instances.
 """
 
+import os
 import json
+from json.decoder import JSONDecodeError
+from ..base_model import BaseModel
 
 
 class FileStorage:
     """FileStorage class.
-    
+
     Private class attributes:
         __file_path: string - path to the JSON file
         __objects: dictionary - empty but will store all objects
-    
+
     Public instance methods:
         all(self): returns the dictionary __objects
         new(self, obj): sets in __objects the obj with key <obj class name>.id
@@ -33,24 +36,33 @@ class FileStorage:
         """sets in __objects the obj with key <obj class name>.id."""
 
         if obj is not None and hasattr(obj, "id"):
-            self.__class__.__objects[f'{obj.__class__.__name__}.{obj.id}'] = obj
+            cls_name = obj.__class__.__name__
+            self.__class__.__objects[f'{cls_name}.{obj.id}'] = obj
 
     def save(self):
         """serializes __objects to the JSON file."""
-        
-        try:
-            with open(self.__class__.__file_path, "w") as f:
-                dict_to_save = {k: v.to_dict() for k, v in self.__class__.__objects.items()}
-                json.dump(dict_to_save, f)
-        except Exception as err:
-            print(err)
-            
+
+        with open(self.__class__.__file_path, "w") as f:
+            dict_of_dicts = {}
+            for key, value in self.__class__.__objects.items():
+                dict_of_dicts[key] = value.to_dict()
+            json.dump(dict_of_dicts, f)
 
     def reload(self):
         """deserializes the JSON file to __objects."""
-        try:
-            with open(self.__class__.__file_path) as f:
-                loaded_dict = json.load(f)
-                self.__class__.__objects = loaded_dict
-        except Exception as err:
-            print(f"Error {err}")
+
+        # We check if file exists otherwise error will be raised
+        if os.path.isfile(self.__class__.__file_path):
+            with open(self.__class__.__file_path, "r+") as f:
+                try:
+                    loaded_dict = json.load(f)
+                    for key, value in loaded_dict.items():
+                        cls_name = value.get("__class__")
+                        obj = eval(cls_name)(**value)
+                        self.new(obj)
+                except JSONDecodeError:
+                    json.dump({}, f)
+
+        else:
+            with open(self.__class__.__file_path, "w") as f:
+                json.dump({}, f)
